graphql/requirements.txt
flask == 3.0.2
pymongo == 4.6.1
strawberry-graphql == 0.219.2